Fixed bug in birthday command
When comparing the birthdays' it included the time, which would basically never match. This has been fixed by resetting the time to midnight when asking for the date today.

Fixed bug in birthday command

When comparing the birthdays' it included the time, which would basically never match. This has been fixed by resetting the time to midnight when asking for the date today.